add incidence matrix class

+++ GraphLib/GraphLib/Graph.cs
@@ -12,7 +12,7 @@ namespace GraphLib
         where TGraphType : GraphType
         where TEdge : IEdge<TVertex>
     {
-        protected const int EMPTY_EDGE = 0; 
+        protected const  double EMPTY_EDGE = 0; 
         public abstract void AddVertex(TVertex vertex);
         public abstract void RemoveVertex(TVertex vertex);
         public abstract void AddEdge(TEdge edge);
